requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
aiofiles==25.1.0
    # via aiogram
aiogram==3.24.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via aiogram
aiosignal==1.4.0
    # via aiohttp
alembic==1.18.1
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   fast-depends
    #   faststream
    #   starlette
asyncio==4.0.0
    # via -r requirements.in
asyncpg==0.31.0
    # via -r requirements.in
attrs==25.4.0
    # via aiohttp
bcrypt==5.0.0
    # via -r requirements.in
certifi==2026.1.4
    # via aiogram
charset-normalizer==3.4.4
    # via -r requirements.in
click==8.3.1
    # via uvicorn
fast-depends[pydantic]==3.0.5
    # via faststream
fastapi==0.128.0
    # via -r requirements.in
faststream==0.6.5
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   yarl
iniconfig==2.3.0
    # via pytest
magic-filter==1.0.12
    # via aiogram
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via pytest
phonenumbers==9.0.21
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   aiogram
    #   fast-depends
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via -r requirements.in
pytest==9.0.2
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
sqlalchemy==2.0.45
    # via
    #   -r requirements.in
    #   alembic
starlette==0.50.0
    # via
    #   -r requirements.in
    #   fastapi
typing-extensions==4.15.0
    # via
    #   aiogram
    #   alembic
    #   fast-depends
    #   fastapi
    #   faststream
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn==0.40.0
    # via -r requirements.in
yarl==1.22.0
    # via aiohttp
